# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StePanel API
  version: 0.6.0
  description: |
    Authenticated server operations, service inventory, and migration API.
    Mutating requests require the `stepanel_session` cookie and the matching
    `stepanel_csrf` value in the `X-CSRF-Token` header or `csrf` form field.
    Responses include an `X-Request-ID` correlation identifier.
servers:
  - url: http://127.0.0.1:8090
security:
  - sessionCookie: []
paths:
...
  /api/sites/environment/{site}:
    get:
      summary: Read site environment metadata with secrets masked
      parameters:
        - {name: site, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Environment metadata}
    put:
      summary: Replace site environment variables
      parameters:
        - {name: site, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, additionalProperties: {type: object}}}}
      responses: {'204': {description: Environment saved}}
    delete:
      summary: Delete all site environment variables
      parameters:
        - {name: site, in: path, required: true, schema: {type: string}}
      responses: {'204': {description: Environment deleted}}
  /api/sites/php/{site}:
    get:
      summary: Inspect a site PHP runtime profile and installed versions
      parameters: [{name: site, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Runtime profile and capability inventory}}
    put:
      summary: Apply a validated per-site PHP-FPM runtime profile
      parameters: [{name: site, in: path, required: true, schema: {type: string}}]
      responses: {'202': {description: FPM profile validated and applied}}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: stepanel_session